Updated web services/API to include JSON
We now receive JSON requests from the front-end UI, process the request,
and return a JSON object if needed. There is still some more stuff to
add, particularly, I don't think the image storage is working entirely,
but troubleshooting with that will go better with a UI. Also, if you are
making a request, you need to make sure your parameters match what I
assumed they would be. Usually the format is all lowercase with spaces
replaced with _